+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Filter: Append a constant 'false' node when an 'inArray' condition with an empty array is appended to a filter tree
+
 ## v0.1.200
 
 * Add enforced filters which, when configured, are always applied for collection GET requests and allows